% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_table.R
\name{na_table}
\alias{na_table}
\title{Generate a Data Frame of Rows With Missing Values}
\usage{
na_table(x)
}
\arguments{
\item{A}{data frame}
}
\value{
A data frame of all rows contain NA in the original.
}
\description{
Generate a Data Frame of Rows With Missing Values
}
\examples{
na_table(mtcarNAs)
}
